requirements.txt
aiohttp
videosdk-agents
videosdk-plugins-openai
videosdk-plugins-google
uvicorn
fastapi
python-dotenv
fastmcp
httpx
PyPDF2

langchain==0.1.20
langchain-core==0.1.52
langchain-openai>=0.0.5
pinecone>=3.0.0
beautifulsoup4>=4.12.3
requests>=2.31.0
tqdm>=4.66.2
